[refactor] : corrected the spelling of FeedbackScreen
