+++ src/navigation/MainStackNavigator.tsx
@@ -10,7 +10,7 @@ const Stack = createNativeStackNavigator();
 import BottomTabsNavigator from "./BottomTabsNavigator";
 import LoginScreen from "../screens/LoginScreen";
 import SpeakersScreen from "../screens/SpeakersScreen";
-import FeedBackScreen from "../screens/FeedbackSreen";
+import FeedBackScreen from "../screens/FeedbackScreen";
 
 import { fonts } from "../assets/fonts/fonts";
 import { TouchableOpacity, StyleSheet , Platform , View , Text, ImageBackground } from "react-native";
